refactor(web,shared,worker): improve code structure and separation of concerns
Decompose large hooks, unify shared stores, and split worker monolith into
focused modules for better maintainability and testability.

Phase 1: Extract page logic hook concerns
- Extract useExchangeFilters from useExchangePageLogic (filtering + distance)
- Extract wizard-steps utils and useValidationSubmit from useValidateGameWizard
- Extract compensation-cache utils and useCompensationData from useCompensationForm

Phase 2: Consolidate stores
- Create shared toast store factory (createToastStore) used by web (5 max) and mobile (3 max)
- Extract calendar-auth service and auth-persistence from 492-line web auth store

Phase 3: Split worker
- Split 1,273-line worker/src/index.ts into route handler modules
- Extract shared middleware (origin validation, CORS, rate limiting)
- Create handlers: robots, version, health, ocr, ojp, ical, proxy
- Router reduced to ~140 lines

Phase 4: Decompose components
- Extract SwipeableCardActions from SwipeableCard (action drawer rendering)
