+++ web-app/src/features/exchanges/hooks/useExchangePageLogic.ts
@@ -3,25 +3,17 @@ import { useState, useCallback, useMemo, useOptimistic } from 'react'
 import { useShallow } from 'zustand/react/shallow'
 
 import type { GameExchange } from '@/api/client'
-import { useCalendarConflicts } from '@/features/assignments/hooks/useCalendarConflicts'
-import {
-  hasMinimumGapFromAssignments,
-  DEFAULT_SAME_LOCATION_DISTANCE_KM,
-} from '@/features/assignments/utils/conflict-detection'
 import { useActiveAssociationCode } from '@/features/auth/hooks/useActiveAssociation'
-import { features } from '@/shared/config/features'
 import { useTour } from '@/shared/hooks/useTour'
-import { useTravelTimeFilter } from '@/shared/hooks/useTravelTimeFilter'
 import { useAuthStore } from '@/shared/stores/auth'
 import { useDemoStore, DEMO_USER_PERSON_IDENTITY } from '@/shared/stores/demo'
 import { useSettingsStore } from '@/shared/stores/settings'
 import { groupByWeek } from '@/shared/utils/date-helpers'
-import { calculateCarDistanceKm } from '@/shared/utils/distance'
-import { extractCoordinates } from '@/shared/utils/geo-location'
 import type { SwipeConfig } from '@/types/swipe'
 
 import { TOUR_DUMMY_EXCHANGE } from '../exchange'
 import { useExchangeActions } from './useExchangeActions'
+import { useExchangeFilters, type ExchangeWithDistance } from './useExchangeFilters'
 import { useGameExchanges, type ExchangeStatus } from './useExchanges'
 import { createExchangeActions } from '../utils/exchange-actions'
 
@@ -54,26 +46,14 @@ export function useExchangePageLogic() {
     },
     [currentUserIdentity]
   )
-  const { userRefereeLevel, userRefereeLevelGradationValue } = useDemoStore(
+  const { userRefereeLevel } = useDemoStore(
     useShallow((state) => ({
       userRefereeLevel: state.userRefereeLevel,
-      userRefereeLevelGradationValue: state.userRefereeLevelGradationValue,
     }))
   )
-  const {
-    homeLocation,
-    distanceFilter,
-    travelTimeFilter,
-    levelFilterEnabled,
-    gameGapFilter,
-    hideOwnExchangesByAssociation,
-  } = useSettingsStore(
+
+  const { hideOwnExchangesByAssociation } = useSettingsStore(
     useShallow((state) => ({
-      homeLocation: state.homeLocation,
-      distanceFilter: state.distanceFilter,
-      travelTimeFilter: state.travelTimeFilter,
-      levelFilterEnabled: state.levelFilterEnabled,
-      gameGapFilter: state.gameGapFilter,
       hideOwnExchangesByAssociation: state.hideOwnExchangesByAssociation,
     }))
   )
@@ -114,161 +94,37 @@ export function useExchangePageLogic() {
   // Show loading when switching associations or when query is loading
   const isLoading = isAssociationSwitching || queryLoading
 
-  // Get calendar assignments for game gap filtering
-  const { assignments: calendarAssignments, hasCalendarCode } = useCalendarConflicts()
+  // Dummy exchange data for tour mode
+  const dummyExchanges = useMemo((): ExchangeWithDistance[] | null => {
+    if (!showDummyData) return null
+    const tourExchange = TOUR_DUMMY_EXCHANGE as unknown as GameExchange
+    return [{ exchange: tourExchange, carDistanceKm: null }]
+  }, [showDummyData])
 
-  // Get travel time data for exchanges
-  const { exchangesWithTravelTime, isAvailable: isTravelTimeAvailable } = useTravelTimeFilter(
-    data ?? null
-  )
-
-  // Build travel time lookup map once for both filtering and rendering
-  const travelTimeMap = useMemo(() => {
-    if (!exchangesWithTravelTime)
-      return new Map<string, { minutes: number | null; isLoading: boolean }>()
-    return new Map(
-      exchangesWithTravelTime.map((e) => [
-        e.item.__identity,
-        { minutes: e.travelTimeMinutes, isLoading: e.isLoading },
-      ])
-    )
-  }, [exchangesWithTravelTime])
-
-  // Calculate car distance for each exchange from user's home location
-  const exchangesWithDistance = useMemo(() => {
-    if (!data) return null
-    if (!homeLocation) return data.map((e) => ({ exchange: e, carDistanceKm: null }))
-
-    return data.map((exchange) => {
-      const geoLocation =
-        exchange.refereeGame?.game?.hall?.primaryPostalAddress?.geographicalLocation
-      const hallCoords = extractCoordinates(geoLocation)
-
-      if (!hallCoords) {
-        return { exchange, carDistanceKm: null }
-      }
-
-      const homeCoords = { latitude: homeLocation.latitude, longitude: homeLocation.longitude }
-      const carDistanceKm = calculateCarDistanceKm(homeCoords, hallCoords)
-
-      return { exchange, carDistanceKm }
-    })
-  }, [data, homeLocation])
-
-  // Filter exchanges by user's referee level and distance when filters are enabled
-  const filteredData = useMemo(() => {
-    // When tour is active (or about to auto-start), show ONLY the dummy exchange
-    // to ensure tour works regardless of whether tabs have real data
-    if (showDummyData) {
-      // Safe cast: TourDummyExchange provides all fields used by ExchangeCard
-      const tourExchange = TOUR_DUMMY_EXCHANGE as unknown as GameExchange
-      return [{ exchange: tourExchange, carDistanceKm: null }]
-    }
-
-    if (!exchangesWithDistance) return null
-
-    let result = exchangesWithDistance
-
-    // Apply level filter (only on "open" tab in demo mode)
-    if (
-      levelFilterEnabled &&
-      statusFilter === 'open' &&
-      isDemoMode &&
-      userRefereeLevelGradationValue !== null
-    ) {
-      result = result.filter(({ exchange }) => {
-        const requiredGradationStr = exchange.requiredRefereeLevelGradationValue
-        // If no gradation value, show the exchange (conservative approach)
-        if (requiredGradationStr === undefined || requiredGradationStr === null) {
-          return true
-        }
-        // Parse string to number for comparison (API returns string)
-        const requiredGradation = Number(requiredGradationStr)
-        if (isNaN(requiredGradation)) {
-          return true
-        }
-        // User can officiate if their level meets or exceeds required level
-        // Lower gradation = higher level, so user.gradation <= required.gradation
-        return userRefereeLevelGradationValue <= requiredGradation
-      })
-    }
-
-    // Apply distance filter (only on "open" tab when home location is set)
-    // Uses car distance for more accurate filtering
-    if (distanceFilter.enabled && statusFilter === 'open' && homeLocation) {
-      result = result.filter(({ carDistanceKm }) => {
-        // If no distance available, include the exchange (conservative)
-        if (carDistanceKm === null) return true
-        return carDistanceKm <= distanceFilter.maxDistanceKm
-      })
-    }
-
-    // Apply travel time filter (only on "open" tab when transport is available)
-    if (
-      travelTimeFilter.enabled &&
-      statusFilter === 'open' &&
-      isTravelTimeAvailable &&
-      travelTimeMap.size > 0
-    ) {
-      result = result.filter(({ exchange }) => {
-        const travelTimeData = travelTimeMap.get(exchange.__identity)
-        // If no travel time available, include the exchange (conservative)
-        if (travelTimeData?.minutes === null || travelTimeData?.minutes === undefined) return true
-        return travelTimeData.minutes <= travelTimeFilter.maxTravelTimeMinutes
-      })
-    }
-
-    // Apply game gap filter (only on "open" tab when calendar data is available)
-    // Uses smart conflict detection: games at nearby venues (<=5km) don't trigger conflicts
-    if (
-      gameGapFilter.enabled &&
-      statusFilter === 'open' &&
-      hasCalendarCode &&
-      calendarAssignments.length > 0
-    ) {
-      result = result.filter(({ exchange }) => {
-        const gameStartTime = exchange.refereeGame?.game?.startingDateTime
-        const geoLocation =
-          exchange.refereeGame?.game?.hall?.primaryPostalAddress?.geographicalLocation
-        const venueCoordinates = extractCoordinates(geoLocation)
-
-        return hasMinimumGapFromAssignments(gameStartTime, calendarAssignments, {
-          minGapMinutes: gameGapFilter.minGapMinutes,
-          venueCoordinates,
-          sameLocationDistanceKm: DEFAULT_SAME_LOCATION_DISTANCE_KM,
-        })
-      })
-    }
-
-    // Hide user's own exchanges in "open" tab when filter is enabled
-    if (hideOwnExchanges && statusFilter === 'open' && currentUserIdentity) {
-      result = result.filter(
-        ({ exchange }) => exchange.submittedByPerson?.__identity !== currentUserIdentity
-      )
-    }
-
-    return result
-  }, [
-    showDummyData,
-    exchangesWithDistance,
+  // Delegate all filtering to the extracted hook
+  const {
+    filteredData,
+    travelTimeMap,
+    homeLocation,
+    isTravelTimeFilterAvailable,
+    isGameGapFilterAvailable,
+    distanceFilter,
+    travelTimeFilter,
+    gameGapFilter,
     levelFilterEnabled,
+  } = useExchangeFilters({
+    data,
     statusFilter,
     isDemoMode,
-    userRefereeLevelGradationValue,
-    distanceFilter.enabled,
-    distanceFilter.maxDistanceKm,
-    homeLocation,
-    travelTimeFilter.enabled,
-    travelTimeFilter.maxTravelTimeMinutes,
-    isTravelTimeAvailable,
-    travelTimeMap,
-    hideOwnExchanges,
     currentUserIdentity,
-    gameGapFilter.enabled,
-    gameGapFilter.minGapMinutes,
-    hasCalendarCode,
-    calendarAssignments,
-  ])
+    hideOwnExchanges,
+    showDummyData,
+    dummyExchanges,
+  })
+
+  // Determine if filters are available
+  const isLevelFilterAvailable = isDemoMode && userRefereeLevel !== null
+  const isDistanceFilterAvailable = homeLocation !== null
 
   // Group exchanges by week for visual separation
   const groupedData = useMemo(() => {
@@ -338,14 +194,6 @@ export function useExchangePageLogic() {
       }
     }
   }, [removeFromExchangeModal.exchange, handleRemoveFromExchange, applyOptimisticRemoval, refetch])
-
-  // Determine if filters are available
-  const isLevelFilterAvailable = isDemoMode && userRefereeLevel !== null
-  const isDistanceFilterAvailable = homeLocation !== null
-  // isTravelTimeAvailable from hook already includes association-specific transport check and homeLocation (features.transport)
-  const isTravelTimeFilterAvailable = features.transport && isTravelTimeAvailable
-  // Game gap filter is available when calendar data exists (demo mode or calendar code)
-  const isGameGapFilterAvailable = hasCalendarCode && calendarAssignments.length > 0
 
   // Handler for pull-to-refresh - wraps refetch in async function
   const handleRefresh = useCallback(async () => {
