+++ worker/src/index.ts
@@ -12,63 +12,22 @@
  * @see https://developers.cloudflare.com/waf/rate-limiting-rules/
  */
 
-import type { HeadersWithCookies } from './types'
+import type { Env } from './types'
+import { securityHeaders } from './middleware'
+import { validateOriginAndPreflight } from './middleware'
 import {
-  AUTH_ENDPOINT,
-  CAPTURE_SESSION_TOKEN_HEADER,
-  CORS_PREFLIGHT_MAX_AGE_SECONDS,
   KILL_SWITCH_RETRY_AFTER_SECONDS,
-  OCR_MAX_FILE_SIZE_BYTES,
-  VOLLEYKIT_USER_AGENT,
-  detectSessionIssue,
-  extractICalCode,
-  hasAuthCredentials,
-  isAllowedOrigin,
-  isAllowedPath,
-  isDynamicContent,
-  isPathSafe,
-  isValidICalCode,
-  noCacheHeaders,
   parseAllowedOrigins,
-  requiresApiPrefix,
-  transformAuthFormData,
   validateAllowedOrigins,
-  // Auth lockout imports
-  type AuthLockoutKV,
-  checkLockoutStatus,
-  clearAuthLockout,
-  getAuthLockoutState,
-  isAuthRequest,
-  isFailedLoginResponse,
-  isSuccessfulLoginResponse,
-  recordFailedAttempt,
 } from './utils'
 
-// OCR configuration constants
-const OCR_ALLOWED_MIME_TYPES = ['image/jpeg', 'image/png', 'image/webp', 'application/pdf'] as const
-const OCR_RATE_LIMIT_RETRY_SECONDS = 60
-
-// OJP configuration constants
-/** Maximum request body size for OJP requests (64 KB) */
-const OJP_MAX_BODY_SIZE_BYTES = 64 * 1024
-
-/**
- * Cloudflare Rate Limiter binding interface.
- * @see https://developers.cloudflare.com/workers/runtime-apis/bindings/rate-limit/
- */
-interface RateLimiter {
-  limit(options: { key: string }): Promise<{ success: boolean }>
-}
-
-interface Env {
-  ALLOWED_ORIGINS: string
-  TARGET_HOST: string
-  RATE_LIMITER: RateLimiter
-  KILL_SWITCH?: string // Set to "true" to disable the proxy
-  MISTRAL_API_KEY?: string // API key for Mistral OCR
-  OJP_API_KEY?: string // API key for Swiss public transport OJP 2.0 API
-  AUTH_LOCKOUT?: AuthLockoutKV // KV namespace for auth lockout state
-}
+import { handleRobots } from './handlers/robots'
+import { handleVersion } from './handlers/version'
+import { handleHealth } from './handlers/health'
+import { handleOcr } from './handlers/ocr'
+import { handleOjp } from './handlers/ojp'
+import { handleICal } from './handlers/ical'
+import { handleProxy } from './handlers/proxy'
 
 function validateEnv(env: Env): void {
   if (!env.ALLOWED_ORIGINS || env.ALLOWED_ORIGINS.trim() === '') {
@@ -84,86 +43,6 @@ function validateEnv(env: Env): void {
   }
 }
 
-/**
- * Security headers to protect against common web vulnerabilities.
- */
-function securityHeaders(): HeadersInit {
-  return {
-    'Content-Security-Policy':
-      "default-src 'self'; script-src 'self'; style-src 'self' 'unsafe-inline'; img-src 'self' data:; connect-src 'self'",
-    'X-Frame-Options': 'SAMEORIGIN',
-    'X-Content-Type-Options': 'nosniff',
-    // X-XSS-Protection set to '0' per MDN recommendation:
-    // '1; mode=block' is deprecated and can introduce timing side-channel attacks.
-    // Modern CSP headers make the XSS auditor redundant.
-    'X-XSS-Protection': '0',
-    'Referrer-Policy': 'strict-origin-when-cross-origin',
-    // HSTS: enforce HTTPS for 1 year including subdomains
-    'Strict-Transport-Security': 'max-age=31536000; includeSubDomains',
-    // Restrict browser features not used by the proxy
-    'Permissions-Policy': 'camera=(), microphone=(), geolocation=(), payment=()',
-    // Prevent cross-origin window references
-    'Cross-Origin-Opener-Policy': 'same-origin',
-  }
-}
-
-function corsHeaders(origin: string): HeadersInit {
-  return {
-    'Access-Control-Allow-Origin': origin,
-    'Access-Control-Allow-Methods': 'GET, POST, PUT, DELETE, OPTIONS',
-    // X-Session-Token is used for iOS Safari PWA cookie relay - see SESSION_TOKEN_HEADER
-    // X-Capture-Session-Token is used to request session token capture from redirect responses
-    // These headers bypass iOS Safari's ITP which blocks third-party cookies in PWA mode
-    'Access-Control-Allow-Headers': `Content-Type, Accept, X-Session-Token, ${CAPTURE_SESSION_TOKEN_HEADER}`,
-    // Expose X-Session-Token so JavaScript can read the session cookie relay
-    'Access-Control-Expose-Headers': 'X-Session-Token, Content-Disposition',
-    'Access-Control-Allow-Credentials': 'true',
-    'Access-Control-Max-Age': String(CORS_PREFLIGHT_MAX_AGE_SECONDS),
-  }
-}
-
-/**
- * Custom header for iOS Safari PWA session cookie relay.
- * iOS Safari's Intelligent Tracking Prevention (ITP) blocks third-party cookies
- * in PWA standalone mode. This header allows the client to:
- * 1. Receive session cookies as a readable header (not blocked by ITP)
- * 2. Store the session token in JavaScript
- * 3. Send it back to the proxy, which converts it to a Cookie header
- */
-const SESSION_TOKEN_HEADER = 'X-Session-Token'
-
-/**
- * Extract session cookie values from Set-Cookie headers.
- * Returns a base64-encoded string of cookie name=value pairs.
- */
-function extractSessionCookies(cookies: string[]): string | null {
-  if (cookies.length === 0) return null
-
-  // Extract just the cookie name=value part (before attributes like Path, Secure, etc.)
-  const cookieValues = cookies
-    .map((cookie) => {
-      const [nameValue] = cookie.split(';')
-      return nameValue?.trim()
-    })
-    .filter(Boolean)
-
-  if (cookieValues.length === 0) return null
-
-  // Join all cookies and base64 encode for safe transport
-  return btoa(cookieValues.join('; '))
-}
-
-/**
- * Decode session token back to cookie string.
- */
-function decodeSessionToken(token: string): string | null {
-  try {
-    return atob(token)
-  } catch {
-    return null
-  }
-}
-
 export default {
   async fetch(request: Request, env: Env): Promise<Response> {
     const url = new URL(request.url)
@@ -171,76 +50,12 @@ export default {
     // Robots.txt - prevent search engine indexing of the proxy
     // This is placed before the kill switch so search engines can always fetch it
     if (url.pathname === '/robots.txt') {
-      return new Response('User-agent: *\nDisallow: /\n', {
-        status: 200,
-        headers: {
-          'Content-Type': 'text/plain',
-          ...securityHeaders(),
-        },
-      })
+      return handleRobots(request, env, url)
     }
 
-    // Version endpoint - returns worker git hash for PWA version tracking.
-    // The web app compares this against its stored worker version to determine
-    // if session tokens need to be invalidated (worker auth logic changed).
-    // This endpoint is public and includes CORS headers for browser access.
+    // Version endpoint - public, before kill switch
     if (url.pathname === '/version') {
-      const origin = request.headers.get('Origin')
-      let allowedOrigins: string[]
-      try {
-        allowedOrigins = parseAllowedOrigins(env.ALLOWED_ORIGINS)
-      } catch {
-        return new Response('Server configuration error', {
-          status: 500,
-          headers: { 'Content-Type': 'text/plain' },
-        })
-      }
-
-      // Check origin for version endpoint (required for CORS)
-      // isAllowedOrigin returns false for null origin, so after this check origin is guaranteed non-null
-      if (!origin || !isAllowedOrigin(origin, allowedOrigins)) {
-        const errorHeaders: HeadersInit = {
-          'Content-Type': 'text/plain',
-          ...securityHeaders(),
-        }
-        if (origin) {
-          Object.assign(errorHeaders, corsHeaders(origin))
-        }
-        return new Response('Forbidden: Origin not allowed', {
-          status: 403,
-          headers: errorHeaders,
-        })
-      }
-
-      // At this point, origin is guaranteed to be a non-null string
-
-      // Handle CORS preflight for version endpoint
-      if (request.method === 'OPTIONS') {
-        return new Response(null, {
-          status: 204,
-          headers: corsHeaders(origin),
-        })
-      }
-
-      // Return worker version info
-      // __WORKER_GIT_HASH__ is injected at deploy time via wrangler --define
-      const workerGitHash = typeof __WORKER_GIT_HASH__ !== 'undefined' ? __WORKER_GIT_HASH__ : 'dev'
-      return new Response(
-        JSON.stringify({
-          workerGitHash,
-          timestamp: new Date().toISOString(),
-        }),
-        {
-          status: 200,
-          headers: {
-            'Content-Type': 'application/json',
-            // Cache for 5 minutes - version doesn't change within a deployment
-            'Cache-Control': 'public, max-age=300',
-            ...corsHeaders(origin),
-            ...securityHeaders(),
-          },
-        }
-      )
+      return handleVersion(request, env, url)
     }
 
     // Kill switch - immediately disable proxy if requested by Swiss Volley
@@ -258,521 +73,19 @@ export default {
       )
     }
 
-    // Health check endpoint - no authentication needed, no rate limiting
-    // Requires CORS headers for browser-based health checks (e.g., OCR availability check)
-    // Verifies Mistral OCR API connectivity by calling /v1/models endpoint
+    // Health check endpoint
     if (url.pathname === '/health') {
-      const origin = request.headers.get('Origin')
-      let allowedOrigins: string[]
-      try {
-        allowedOrigins = parseAllowedOrigins(env.ALLOWED_ORIGINS)
-      } catch {
-        return new Response('Server configuration error', {
-          status: 500,
-          headers: { 'Content-Type': 'text/plain' },
-        })
-      }
-
-      // Check origin for health endpoint (required for CORS)
-      // Include CORS headers even in error responses so browsers can read the error
-      if (!isAllowedOrigin(origin, allowedOrigins)) {
-        const errorHeaders: HeadersInit = {
-          'Content-Type': 'text/plain',
-          ...securityHeaders(),
-        }
-        // If origin was provided, include CORS headers so browser can read the error
-        if (origin) {
-          Object.assign(errorHeaders, corsHeaders(origin))
-        }
-        return new Response('Forbidden: Origin not allowed', {
-          status: 403,
-          headers: errorHeaders,
-        })
-      }
-
-      // Handle CORS preflight for health check
-      if (request.method === 'OPTIONS') {
-        return new Response(null, {
-          status: 204,
-          headers: corsHeaders(origin!),
-        })
-      }
-
-      // Check Mistral API connectivity if API key is configured
-      // Uses /v1/models endpoint as a lightweight health check (no token consumption)
-      // Timeout after 5 seconds to prevent health check from hanging
-      const HEALTH_CHECK_TIMEOUT_MS = 5000
-      let mistralStatus: 'ok' | 'not_configured' | 'error' = 'not_configured'
-      let mistralError: string | undefined
-
-      if (env.MISTRAL_API_KEY) {
-        const controller = new AbortController()
-        const timeoutId = setTimeout(() => controller.abort(), HEALTH_CHECK_TIMEOUT_MS)
-
-        try {
-          const mistralResponse = await fetch('https://api.mistral.ai/v1/models', {
-            method: 'GET',
-            headers: {
-              Authorization: `Bearer ${env.MISTRAL_API_KEY}`,
-            },
-            signal: controller.signal,
-          })
-
-          if (mistralResponse.ok) {
-            mistralStatus = 'ok'
-          } else {
-            mistralStatus = 'error'
-            mistralError =
-              mistralResponse.status === 401
-                ? 'Invalid API key'
-                : `API returned ${mistralResponse.status}`
-          }
-        } catch (error) {
-          mistralStatus = 'error'
-          if (error instanceof Error && error.name === 'AbortError') {
-            mistralError = 'Health check timed out'
-          } else {
-            mistralError = error instanceof Error ? error.message : 'Connection failed'
-          }
-        } finally {
-          clearTimeout(timeoutId)
-        }
-      }
-
-      const overallStatus = mistralStatus === 'ok' ? 'ok' : 'degraded'
-
-      return new Response(
-        JSON.stringify({
-          status: overallStatus,
-          timestamp: new Date().toISOString(),
-          services: {
-            proxy: 'ok',
-            mistral_ocr: mistralStatus,
-            ...(mistralError && { mistral_ocr_error: mistralError }),
-          },
-        }),
-        {
-          status: overallStatus === 'ok' ? 200 : 503,
-          headers: {
-            'Content-Type': 'application/json',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        }
-      )
+      return handleHealth(request, env, url)
     }
 
     // OCR endpoint - proxy to Mistral OCR API
-    // Accepts POST with image data and returns OCR results
     if (url.pathname === '/ocr') {
-      const origin = request.headers.get('Origin')
-      let allowedOrigins: string[]
-      try {
-        allowedOrigins = parseAllowedOrigins(env.ALLOWED_ORIGINS)
-      } catch {
-        return new Response('Server configuration error', {
-          status: 500,
-          headers: { 'Content-Type': 'text/plain' },
-        })
-      }
-
-      // Check origin for OCR endpoint
-      // Include CORS headers even in error responses so browsers can read the error
-      if (!isAllowedOrigin(origin, allowedOrigins)) {
-        const errorHeaders: HeadersInit = {
-          'Content-Type': 'text/plain',
-          ...securityHeaders(),
-        }
-        // If origin was provided, include CORS headers so browser can read the error
-        if (origin) {
-          Object.assign(errorHeaders, corsHeaders(origin))
-        }
-        return new Response('Forbidden: Origin not allowed', {
-          status: 403,
-          headers: errorHeaders,
-        })
-      }
-
-      // Handle CORS preflight for OCR
-      if (request.method === 'OPTIONS') {
-        return new Response(null, {
-          status: 204,
-          headers: corsHeaders(origin!),
-        })
-      }
-
-      // Rate limiting for OCR endpoint
-      // CF-Connecting-IP is always present in production (set by Cloudflare edge).
-      // Skip rate limiting if absent (test/dev only) to avoid a shared 'unknown' bucket.
-      if (env.RATE_LIMITER) {
-        const clientIP = request.headers.get('CF-Connecting-IP')
-        if (!clientIP) {
-          console.warn('CF-Connecting-IP header missing — skipping rate limit')
-        }
-        const { success } = clientIP
-          ? await env.RATE_LIMITER.limit({ key: clientIP })
-          : { success: true }
-
-        if (!success) {
-          return new Response(JSON.stringify({ error: 'Too many requests' }), {
-            status: 429,
-            headers: {
-              'Content-Type': 'application/json',
-              'Retry-After': String(OCR_RATE_LIMIT_RETRY_SECONDS),
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-      }
-
-      // Only allow POST for OCR
-      if (request.method !== 'POST') {
-        return new Response('Method Not Allowed', {
-          status: 405,
-          headers: {
-            'Content-Type': 'text/plain',
-            Allow: 'POST',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
-
-      // Check if Mistral API key is configured
-      if (!env.MISTRAL_API_KEY) {
-        return new Response(JSON.stringify({ error: 'OCR service not configured' }), {
-          status: 503,
-          headers: {
-            'Content-Type': 'application/json',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
-
-      try {
-        // Parse the incoming request - expects multipart form data with 'image' field
-        const formData = await request.formData()
-        const imageFile = formData.get('image') as File | null
-
-        if (!imageFile) {
-          return new Response(JSON.stringify({ error: "Missing 'image' field in form data" }), {
-            status: 400,
-            headers: {
-              'Content-Type': 'application/json',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-
-        // Validate file type
-        if (
-          !OCR_ALLOWED_MIME_TYPES.includes(
-            imageFile.type as (typeof OCR_ALLOWED_MIME_TYPES)[number]
-          )
-        ) {
-          return new Response(
-            JSON.stringify({
-              error: `Unsupported file type: ${imageFile.type}. Allowed: ${OCR_ALLOWED_MIME_TYPES.join(', ')}`,
-            }),
-            {
-              status: 400,
-              headers: {
-                'Content-Type': 'application/json',
-                ...corsHeaders(origin!),
-                ...securityHeaders(),
-              },
-            }
-          )
-        }
-
-        // Validate file size
-        if (imageFile.size > OCR_MAX_FILE_SIZE_BYTES) {
-          return new Response(
-            JSON.stringify({
-              error: `File too large: ${(imageFile.size / 1024 / 1024).toFixed(1)}MB. Maximum: 50MB`,
-            }),
-            {
-              status: 400,
-              headers: {
-                'Content-Type': 'application/json',
-                ...corsHeaders(origin!),
-                ...securityHeaders(),
-              },
-            }
-          )
-        }
-
-        // Convert image to base64 for Mistral API
-        // Use chunked approach to avoid stack overflow with large files
-        const imageBuffer = await imageFile.arrayBuffer()
-        const bytes = new Uint8Array(imageBuffer)
-        const CHUNK_SIZE = 8192
-        let base64Image = ''
-        for (let i = 0; i < bytes.length; i += CHUNK_SIZE) {
-          const chunk = bytes.subarray(i, i + CHUNK_SIZE)
-          base64Image += String.fromCharCode(...chunk)
-        }
-        base64Image = btoa(base64Image)
-        const dataUrl = `data:${imageFile.type};base64,${base64Image}`
-
-        // Call Mistral OCR API
-        const mistralResponse = await fetch('https://api.mistral.ai/v1/ocr', {
-          method: 'POST',
-          headers: {
-            Authorization: `Bearer ${env.MISTRAL_API_KEY}`,
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            model: 'mistral-ocr-latest',
-            document: {
-              type: 'image_url',
-              image_url: {
-                url: dataUrl,
-              },
-            },
-            include_image_base64: false,
-            // Enable HTML table formatting for structured table extraction
-            // Scoresheets contain player lists and scores that benefit from table structure
-            table_format: 'html',
-          }),
-        })
-
-        if (!mistralResponse.ok) {
-          const errorText = await mistralResponse.text()
-          console.error('Mistral OCR error:', mistralResponse.status, errorText)
-
-          // Return appropriate error based on status
-          if (mistralResponse.status === 401) {
-            return new Response(JSON.stringify({ error: 'OCR service authentication failed' }), {
-              status: 503,
-              headers: {
-                'Content-Type': 'application/json',
-                ...corsHeaders(origin!),
-                ...securityHeaders(),
-              },
-            })
-          }
-
-          if (mistralResponse.status === 429) {
-            return new Response(JSON.stringify({ error: 'OCR service rate limit exceeded' }), {
-              status: 429,
-              headers: {
-                'Content-Type': 'application/json',
-                'Retry-After': String(OCR_RATE_LIMIT_RETRY_SECONDS),
-                ...corsHeaders(origin!),
-                ...securityHeaders(),
-              },
-            })
-          }
-
-          return new Response(JSON.stringify({ error: 'OCR processing failed' }), {
-            status: 502,
-            headers: {
-              'Content-Type': 'application/json',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-
-        // Return the Mistral OCR response
-        const ocrResult = await mistralResponse.json()
-        return new Response(JSON.stringify(ocrResult), {
-          status: 200,
-          headers: {
-            'Content-Type': 'application/json',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      } catch (error) {
-        console.error('OCR proxy error:', error)
-        return new Response(
-          JSON.stringify({ error: 'Internal server error during OCR processing' }),
-          {
-            status: 500,
-            headers: {
-              'Content-Type': 'application/json',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          }
-        )
-      }
+      return handleOcr(request, env, url)
     }
 
-    // OJP proxy endpoint - proxies requests to Swiss public transport OJP 2.0 API
-    // Keeps the API key server-side instead of exposing it in the client bundle
+    // OJP proxy endpoint - Swiss public transport OJP 2.0 API
     if (url.pathname === '/ojp') {
-      const origin = request.headers.get('Origin')
-      let allowedOrigins: string[]
-      try {
-        allowedOrigins = parseAllowedOrigins(env.ALLOWED_ORIGINS)
-      } catch {
-        return new Response('Server configuration error', {
-          status: 500,
-          headers: { 'Content-Type': 'text/plain' },
-        })
-      }
-
-      if (!isAllowedOrigin(origin, allowedOrigins)) {
-        const errorHeaders: HeadersInit = {
-          'Content-Type': 'text/plain',
-          ...securityHeaders(),
-        }
-        if (origin) {
-          Object.assign(errorHeaders, corsHeaders(origin))
-        }
-        return new Response('Forbidden: Origin not allowed', {
-          status: 403,
-          headers: errorHeaders,
-        })
-      }
-
-      // Handle CORS preflight for OJP
-      if (request.method === 'OPTIONS') {
-        return new Response(null, {
-          status: 204,
-          headers: corsHeaders(origin!),
-        })
-      }
-
-      // Rate limiting for OJP endpoint
-      // CF-Connecting-IP is always present in production (set by Cloudflare edge).
-      // Skip rate limiting if absent (test/dev only) to avoid a shared 'unknown' bucket.
-      if (env.RATE_LIMITER) {
-        const clientIP = request.headers.get('CF-Connecting-IP')
-        if (!clientIP) {
-          console.warn('CF-Connecting-IP header missing — skipping rate limit')
-        }
-        const { success } = clientIP
-          ? await env.RATE_LIMITER.limit({ key: clientIP })
-          : { success: true }
-
-        if (!success) {
-          return new Response(JSON.stringify({ error: 'Too many requests' }), {
-            status: 429,
-            headers: {
-              'Content-Type': 'application/json',
-              'Retry-After': '60',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-      }
-
-      // Only allow POST for OJP
-      if (request.method !== 'POST') {
-        return new Response('Method Not Allowed', {
-          status: 405,
-          headers: {
-            'Content-Type': 'text/plain',
-            Allow: 'POST',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
-
-      // Check if OJP API key is configured
-      if (!env.OJP_API_KEY) {
-        return new Response(JSON.stringify({ error: 'OJP service not configured' }), {
-          status: 503,
-          headers: {
-            'Content-Type': 'application/json',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
-
-      try {
-        // Forward the request body to OJP API with server-side auth
-        const ojpBody = await request.text()
-
-        // Validate body size to prevent abuse (OJP XML requests are typically a few KB)
-        if (ojpBody.length > OJP_MAX_BODY_SIZE_BYTES) {
-          return new Response(JSON.stringify({ error: 'Request body too large' }), {
-            status: 413,
-            headers: {
-              'Content-Type': 'application/json',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-
-        const ojpResponse = await fetch('https://api.opentransportdata.swiss/ojp20', {
-          method: 'POST',
-          headers: {
-            Authorization: `Bearer ${env.OJP_API_KEY}`,
-            'Content-Type': request.headers.get('Content-Type') || 'application/xml',
-          },
-          body: ojpBody,
-        })
-
-        if (!ojpResponse.ok) {
-          const errorText = await ojpResponse.text()
-          console.error('OJP API error:', ojpResponse.status, errorText)
-
-          if (ojpResponse.status === 401) {
-            return new Response(JSON.stringify({ error: 'OJP service authentication failed' }), {
-              status: 503,
-              headers: {
-                'Content-Type': 'application/json',
-                ...corsHeaders(origin!),
-                ...securityHeaders(),
-              },
-            })
-          }
-
-          if (ojpResponse.status === 429) {
-            return new Response(JSON.stringify({ error: 'OJP service rate limit exceeded' }), {
-              status: 429,
-              headers: {
-                'Content-Type': 'application/json',
-                'Retry-After': '60',
-                ...corsHeaders(origin!),
-                ...securityHeaders(),
-              },
-            })
-          }
-
-          return new Response(JSON.stringify({ error: 'OJP request failed' }), {
-            status: 502,
-            headers: {
-              'Content-Type': 'application/json',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-
-        // Return the OJP API response
-        const ojpResult = await ojpResponse.text()
-        return new Response(ojpResult, {
-          status: 200,
-          headers: {
-            'Content-Type': ojpResponse.headers.get('Content-Type') || 'application/xml',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      } catch (error) {
-        console.error('OJP proxy error:', error)
-        return new Response(JSON.stringify({ error: 'Internal server error during OJP request' }), {
-          status: 500,
-          headers: {
-            'Content-Type': 'application/json',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
+      return handleOjp(request, env, url)
     }
 
     // Rate limiting check (100 requests per minute per IP)
@@ -814,460 +127,16 @@ export default {
       })
     }
 
-    const origin = request.headers.get('Origin')
-
-    // Check if origin is allowed
-    // Include CORS headers even in error responses so browsers can read the error
-    if (!isAllowedOrigin(origin, allowedOrigins)) {
-      const errorHeaders: HeadersInit = {
-        'Content-Type': 'text/plain',
-        ...securityHeaders(),
-      }
-      // If origin was provided, include CORS headers so browser can read the error
-      if (origin) {
-        Object.assign(errorHeaders, corsHeaders(origin))
-      }
-      return new Response('Forbidden: Origin not allowed', {
-        status: 403,
-        headers: errorHeaders,
-      })
-    }
-
-    // Handle CORS preflight
-    if (request.method === 'OPTIONS') {
-      return new Response(null, {
-        status: 204,
-        headers: corsHeaders(origin!),
-      })
-    }
+    // Validate origin and handle CORS preflight
+    const check = validateOriginAndPreflight(request, env)
+    if (check.response) return check.response
+    const origin = check.origin
 
     // Handle iCal proxy route: GET/HEAD /iCal/referee/:code
-    // This is a public endpoint that proxies to volleymanager's iCal feed
-    // HEAD is used by the client to validate calendar codes exist without downloading content
-    const iCalCode = extractICalCode(url.pathname)
-    if (iCalCode !== null) {
-      // Only allow GET and HEAD requests for iCal
-      if (request.method !== 'GET' && request.method !== 'HEAD') {
-        return new Response('Method Not Allowed', {
-          status: 405,
-          headers: {
-            'Content-Type': 'text/plain',
-            Allow: 'GET, HEAD',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
+    const iCalResponse = await handleICal(request, env, url, origin)
+    if (iCalResponse) return iCalResponse
 
-      // Validate code format (6 alphanumeric characters)
-      if (!isValidICalCode(iCalCode)) {
-        return new Response('Bad Request: Invalid calendar code format', {
-          status: 400,
-          headers: {
-            'Content-Type': 'text/plain',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
-
-      // Proxy to volleymanager iCal endpoint
-      // Note: We always use GET for the upstream request, even for HEAD requests from clients.
-      // This is because the upstream server may not support HEAD, and we need to verify the
-      // calendar exists. For HEAD requests, we simply discard the response body.
-      const iCalTargetUrl = `${env.TARGET_HOST}/indoor/iCal/referee/${iCalCode}`
-
-      try {
-        const iCalResponse = await fetch(iCalTargetUrl, {
-          method: 'GET',
-          headers: {
-            'User-Agent': VOLLEYKIT_USER_AGENT,
-            Accept: 'text/calendar',
-          },
-        })
-
-        // Pass through 404 if not found
-        if (iCalResponse.status === 404) {
-          return new Response('Not Found: Calendar not found', {
-            status: 404,
-            headers: {
-              'Content-Type': 'text/plain',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-
-        // Handle other error responses from upstream
-        if (!iCalResponse.ok) {
-          console.error('iCal upstream error:', iCalResponse.status, iCalResponse.statusText)
-          return new Response('Bad Gateway: Upstream error', {
-            status: 502,
-            headers: {
-              'Content-Type': 'text/plain',
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          })
-        }
-
-        // Return the iCal data with proper headers
-        // For HEAD requests, return empty body (used for validation)
-        const iCalBody = request.method === 'HEAD' ? null : await iCalResponse.text()
-        return new Response(iCalBody, {
-          status: 200,
-          headers: {
-            'Content-Type': 'text/calendar; charset=utf-8',
-            'Cache-Control': 'public, max-age=300', // 5 minutes
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      } catch (error) {
-        console.error('iCal proxy error:', error)
-        return new Response('Bad Gateway: Unable to reach calendar service', {
-          status: 502,
-          headers: {
-            'Content-Type': 'text/plain',
-            ...corsHeaders(origin!),
-            ...securityHeaders(),
-          },
-        })
-      }
-    }
-
-    // Validate pathname for security (path traversal prevention)
-    if (!isPathSafe(url.pathname)) {
-      return new Response('Bad Request: Invalid path', {
-        status: 400,
-        headers: {
-          'Content-Type': 'text/plain',
-          ...corsHeaders(origin!),
-          ...securityHeaders(),
-        },
-      })
-    }
-
-    // Only proxy allowed paths
-    if (!isAllowedPath(url.pathname)) {
-      return new Response('Not Found: Path not proxied', {
-        status: 404,
-        headers: {
-          'Content-Type': 'text/plain',
-          ...corsHeaders(origin!),
-          ...securityHeaders(),
-        },
-      })
-    }
-
-    // Auth lockout check - block auth requests from locked-out IPs
-    // This is tamper-proof as the state is stored server-side in KV
-    // CF-Connecting-IP is always present in production (set by Cloudflare edge).
-    // If absent (test/dev), auth lockout is skipped to avoid a shared 'unknown' bucket.
-    const clientIP = request.headers.get('CF-Connecting-IP')
-    const isAuthReq = isAuthRequest(url.pathname, request.method)
-
-    if (isAuthReq && env.AUTH_LOCKOUT && clientIP) {
-      const lockoutState = await getAuthLockoutState(env.AUTH_LOCKOUT, clientIP)
-      const lockoutStatus = checkLockoutStatus(lockoutState)
-
-      if (lockoutStatus.isLocked) {
-        return new Response(
-          JSON.stringify({
-            error: 'Too many failed login attempts',
-            lockedUntil: lockoutStatus.remainingSeconds,
-            message: `Account locked. Try again in ${lockoutStatus.remainingSeconds} seconds.`,
-          }),
-          {
-            status: 423, // 423 Locked
-            headers: {
-              'Content-Type': 'application/json',
-              'Retry-After': String(lockoutStatus.remainingSeconds),
-              ...corsHeaders(origin!),
-              ...securityHeaders(),
-            },
-          }
-        )
-      }
-    }
-
-    // Build target URL
-    // Extract the raw path + search from request URL to preserve URL encoding
-    // This is critical for backslash (%5c) encoding required by TYPO3 Neos/Flow
-    const requestUrlStr = request.url
-    const pathStart = requestUrlStr.indexOf('/', requestUrlStr.indexOf('://') + 3)
-    let rawPathAndSearch = pathStart >= 0 ? requestUrlStr.substring(pathStart) : '/'
-
-    // iOS Safari workaround: Detect requests to /login with auth credentials
-    // iOS Safari can trigger form submission with incorrect method (GET instead of POST)
-    // or send to wrong endpoint. This rewrites such requests to the correct auth endpoint.
-    let requestBody: string | null = null
-    let methodOverride: string | null = null
-    if (url.pathname === '/login' && (request.method === 'POST' || request.method === 'GET')) {
-      // Check for auth credentials in body (POST) or query params (GET)
-      if (
-        request.method === 'POST' &&
-        request.headers.get('Content-Type')?.includes('application/x-www-form-urlencoded')
-      ) {
-        // Clone the request to read the body (body can only be read once)
-        const clonedRequest = request.clone()
-        requestBody = await clonedRequest.text()
-
-        if (hasAuthCredentials(requestBody)) {
-          // Rewrite to correct auth endpoint (iOS Safari workaround)
-          rawPathAndSearch = AUTH_ENDPOINT
-          requestBody = transformAuthFormData(requestBody)
-        }
-      } else if (request.method === 'GET' && url.search) {
-        // iOS Safari bug: form may submit as GET with credentials in query string
-        // or with postData attached to GET (invalid HTTP, but Safari does it)
-        const queryBody = url.search.slice(1) // Remove leading '?'
-        if (hasAuthCredentials(queryBody)) {
-          // Convert GET to POST and rewrite to correct auth endpoint (iOS Safari workaround)
-          rawPathAndSearch = AUTH_ENDPOINT
-          requestBody = transformAuthFormData(queryBody)
-          methodOverride = 'POST'
-        }
-      }
-    }
-
-    // Prepend /api/ prefix for API endpoints
-    // Note: We check the original pathname, not rawPathAndSearch, because iOS Safari
-    // workaround may have rewritten rawPathAndSearch to AUTH_ENDPOINT
-    const needsApiPrefix = requiresApiPrefix(url.pathname)
-    const finalPath = needsApiPrefix ? `/api${rawPathAndSearch}` : rawPathAndSearch
-    const targetUrl = new URL(finalPath, env.TARGET_HOST)
-
-    // Forward the request
-    // If we already read the body for iOS workaround detection, use it directly
-    // methodOverride is used when converting GET to POST for iOS auth workaround
-    const proxyRequest = new Request(targetUrl.toString(), {
-      method: methodOverride ?? request.method,
-      headers: request.headers,
-      body: requestBody ?? request.body,
-      redirect: 'manual', // Handle redirects manually to preserve cookies
-    })
-
-    // If we're overriding to POST, ensure Content-Type is set
-    if (methodOverride === 'POST' && requestBody) {
-      proxyRequest.headers.set('Content-Type', 'application/x-www-form-urlencoded')
-    }
-
-    // Remove headers that shouldn't be forwarded
-    proxyRequest.headers.delete('Host')
-    const targetHostUrl = new URL(env.TARGET_HOST)
-    const targetHost = targetHostUrl.host
-    const targetOrigin = targetHostUrl.origin
-    proxyRequest.headers.set('Host', targetHost)
-
-    // Rewrite Origin and Referer to match target host
-    // The upstream Neos Flow server validates these for CSRF protection
-    proxyRequest.headers.set('Origin', targetOrigin)
-
-    const originalReferer = proxyRequest.headers.get('Referer')
-    if (originalReferer) {
-      try {
-        const refererUrl = new URL(originalReferer)
-        refererUrl.protocol = targetHostUrl.protocol
-        refererUrl.host = targetHost
-        // Strip query parameters from Referer - the upstream server may reject
-        // requests with unknown query params like ?debug
-        refererUrl.search = ''
-        proxyRequest.headers.set('Referer', refererUrl.toString())
-      } catch {
-        // If Referer is malformed, set a safe default
-        proxyRequest.headers.set('Referer', targetOrigin + '/')
-      }
-    }
-
-    proxyRequest.headers.set('User-Agent', VOLLEYKIT_USER_AGENT)
-
-    // iOS Safari PWA cookie relay: convert X-Session-Token header to Cookie
-    // This bypasses iOS Safari's ITP which blocks third-party cookies in PWA mode
-    const sessionToken = request.headers.get(SESSION_TOKEN_HEADER)
-    if (sessionToken) {
-      const cookieValue = decodeSessionToken(sessionToken)
-      if (cookieValue) {
-        // Merge with any existing cookies (from browsers that do support cookies)
-        const existingCookies = proxyRequest.headers.get('Cookie')
-        const mergedCookies = existingCookies ? `${existingCookies}; ${cookieValue}` : cookieValue
-        proxyRequest.headers.set('Cookie', mergedCookies)
-      }
-    }
-
-    // Record request start time for latency tracking
-    const requestStartTime = Date.now()
-
-    try {
-      const response = await fetch(proxyRequest)
-
-      // Track auth attempts for lockout
-      // Must clone response to read body without consuming it
-      let responseBodyForCheck: string | undefined
-      if (isAuthReq && env.AUTH_LOCKOUT && clientIP) {
-        // Clone response to read body for auth result detection
-        const clonedResponse = response.clone()
-        try {
-          responseBodyForCheck = await clonedResponse.text()
-        } catch {
-          // If we can't read the body, continue without it
-        }
-
-        const isSuccess = isSuccessfulLoginResponse(response)
-        const isFailed = isFailedLoginResponse(response, responseBodyForCheck)
-
-        if (isSuccess) {
-          // Clear lockout on successful login
-          await clearAuthLockout(env.AUTH_LOCKOUT, clientIP)
-        } else if (isFailed) {
-          // Record failed attempt
-          await recordFailedAttempt(env.AUTH_LOCKOUT, clientIP)
-        }
-      }
-
-      // Create response with CORS headers
-      const responseHeaders = new Headers(response.headers)
-
-      // Add CORS and security headers
-      Object.entries({ ...corsHeaders(origin!), ...securityHeaders() }).forEach(([key, value]) => {
-        responseHeaders.set(key, value)
-      })
-
-      // Apply cache control based on content type
-      // Dynamic content (HTML, JSON) should never be cached to prevent stale data
-      const contentType = response.headers.get('Content-Type')
-      if (isDynamicContent(contentType)) {
-        // Remove any upstream cache headers that could cause stale data
-        responseHeaders.delete('ETag')
-        responseHeaders.delete('Last-Modified')
-
-        // Apply strict no-cache headers
-        Object.entries(noCacheHeaders()).forEach(([key, value]) => {
-          responseHeaders.set(key, value)
-        })
-      }
-
-      // Add proxy timing header for debugging stale data issues
-      // Format: timestamp when proxy processed request, latency in ms
-      const latencyMs = Date.now() - requestStartTime
-      responseHeaders.set(
-        'X-Proxy-Timestamp',
-        `${new Date().toISOString()}; latency=${latencyMs}ms`
-      )
-
-      // Detect potential session issues for logging/debugging
-      // This helps identify when cached responses might be from expired sessions
-      if (detectSessionIssue(response)) {
-        responseHeaders.set('X-Proxy-Session-Warning', 'potential-session-issue')
-      }
-
-      // Handle Set-Cookie headers for cross-origin
-      // Must set SameSite=None; Secure for cross-origin cookies to work
-      // Use getSetCookie() to handle multiple cookies (get() only returns first)
-      const cookies = (response.headers as HeadersWithCookies).getSetCookie()
-      if (cookies.length > 0) {
-        // First, delete any existing Set-Cookie headers
-        responseHeaders.delete('Set-Cookie')
-
-        for (const cookie of cookies) {
-          // Remove Domain, Secure, SameSite, and Partitioned - we'll add them back explicitly
-          // This ensures consistent behavior regardless of original cookie format
-          const modifiedCookie = cookie
-            .replace(/Domain=[^;]+;?\s*/gi, '')
-            .replace(/;\s*Secure\s*(;|$)/gi, '$1')
-            .replace(/SameSite=[^;]+;?\s*/gi, '')
-            .replace(/;\s*Partitioned\s*(;|$)/gi, '$1')
-
-          // Add SameSite=None, Secure, and Partitioned for cross-origin compatibility
-          // Partitioned (CHIPS) is required for iOS Safari's ITP to allow third-party cookies
-          responseHeaders.append(
-            'Set-Cookie',
-            `${modifiedCookie}; SameSite=None; Secure; Partitioned`
-          )
-        }
-
-        // iOS Safari PWA cookie relay: also send cookies as X-Session-Token header
-        // This allows JavaScript to receive cookies even when ITP blocks Set-Cookie
-        const extractedSessionToken = extractSessionCookies(cookies)
-        if (extractedSessionToken) {
-          responseHeaders.set(SESSION_TOKEN_HEADER, extractedSessionToken)
-        }
-      }
-
-      // Handle redirects
-      if (response.status >= 300 && response.status < 400) {
-        const location = response.headers.get('Location')
-        if (location) {
-          // Rewrite redirect to go through proxy
-          const redirectUrl = new URL(location, env.TARGET_HOST)
-          if (redirectUrl.host === new URL(env.TARGET_HOST).host) {
-            // Internal redirect - rewrite to proxy
-            const proxyRedirect = new URL(request.url)
-            proxyRedirect.pathname = redirectUrl.pathname
-            proxyRedirect.search = redirectUrl.search
-
-            // iOS Safari PWA fix: Convert redirects to 200 + JSON response in two cases:
-            //
-            // 1. POST auth requests - Always convert because iOS Safari PWA doesn't
-            //    properly store cookies from redirect responses
-            //
-            // 2. Session capture requests - When client sends X-Capture-Session-Token header
-            //    AND there's a session token to relay. We require both conditions because:
-            //    - wantsSessionCapture alone would unnecessarily convert redirects that have
-            //      no session token (e.g., if upstream didn't set a cookie)
-            //    - hasSessionToken alone would convert all redirects with tokens, even when
-            //      the client doesn't need special handling
-            //    This allows capturing session tokens from redirect responses (which would
-            //    otherwise be opaqueredirect with hidden headers when using redirect: 'manual')
-            //
-            // When using redirect: "manual", browsers return "opaqueredirect" which hides
-            // all headers including Set-Cookie from JavaScript. By returning 200 with JSON,
-            // the browser processes Set-Cookie normally and the client can read the session.
-            const wantsSessionCapture = request.headers.get(CAPTURE_SESSION_TOKEN_HEADER) === 'true'
-            const hasSessionToken = responseHeaders.has(SESSION_TOKEN_HEADER)
-            const shouldConvertToJson =
-              (isAuthReq && request.method === 'POST') || (wantsSessionCapture && hasSessionToken)
-
-            if (shouldConvertToJson) {
-              const isSuccess =
-                isAuthReq && request.method === 'POST' ? isSuccessfulLoginResponse(response) : true // Session capture requests are informational
-              responseHeaders.set('Content-Type', 'application/json')
-              responseHeaders.delete('Location') // Remove redirect header
-              return new Response(
-                JSON.stringify({
-                  success: isSuccess,
-                  redirectUrl: proxyRedirect.toString(),
-                  // Include session token in body for redundancy
-                  // (also available in X-Session-Token header)
-                  sessionCaptured: hasSessionToken,
-                }),
-                {
-                  status: 200,
-                  headers: responseHeaders,
-                }
-              )
-            }
-
-            responseHeaders.set('Location', proxyRedirect.toString())
-          }
-        }
-      }
-
-      return new Response(response.body, {
-        status: response.status,
-        statusText: response.statusText,
-        headers: responseHeaders,
-      })
-    } catch (error) {
-      console.error('Proxy error:', error)
-      return new Response('Proxy Error', {
-        status: 502,
-        headers: {
-          'Content-Type': 'text/plain',
-          ...corsHeaders(origin!),
-          ...securityHeaders(),
-        },
-      })
-    }
+    // Main CORS proxy - forward to volleymanager
+    return handleProxy(request, env, url, origin)
   },
 }
